Fixed warning: the format specifier %.*s requires int (Fixes issue 2). Thanks dreamlax :)
http://www.reddit.com/r/C_Programming/comments/duekp/an_exception_handling_framework_for_ansi_c/c131ef3

+++ src/e4c.c
@@ -991,7 +991,7 @@ static E4C_INLINE e4c_exception _e4c_new_exception(const e4c_exception * prototy
 
 	exception.name			= prototype->name;
 
-	(void)sprintf(exception.message, "%.*s", sizeof(exception.message) - 1, message != NULL ? message : prototype->message);
+	(void)sprintf(exception.message, "%.*s", (int)sizeof(exception.message) - 1, message != NULL ? message : prototype->message);
 
 	exception.super			= prototype->super;
 	exception.file			= file;
